+++ src/initial_conditions/InitialConditions.jl
@@ -9,8 +9,11 @@ import ..Microphysics0Moment
 import ..Microphysics1Moment
 import ..PrescribedSurfaceTemperature
 import ..PrognosticSurfaceTemperature
+import ..ᶜinterp
+import ..ᶠinterp
 import ..C3
 import ..C12
+import ..compute_kinetic!
 import ..PrognosticEDMFX
 import ..DiagnosticEDMFX
 import ..n_mass_flux_subdomains
@@ -30,6 +33,7 @@ import SciMLBase
 import Interpolations as Intp
 import NCDatasets as NC
 import Statistics: mean
+import ClimaUtilities.SpaceVaryingInputs
 
 include("local_state.jl")
 include("atmos_state.jl")